+++ CarDataProject/CarDataProject/Objects/Car.cs
@@ -6,11 +6,11 @@ using System.Threading.Tasks;
 
 namespace CarDataProject {
     class Car {
-        public Car(Int64 carid) {
+        public Car(Int16 carid) {
             this.carId = carid;
         }
 
-        public Int64 carId { get; set; }
+        public Int16 carId { get; set; }
         public List<Trip> allTrips { get; set; }
 
 
@@ -20,12 +20,12 @@ namespace CarDataProject {
 
         }
 
-        public void UpdateCarWithTripIds(Int64 carid) {
+        public void UpdateCarWithTripIds(Int16 carid) {
             DBController dbc = new DBController();
 
             List<Trip> myTrips = TripCalculator.CalculateTripsByCarId(1);
             for (int i = 1; i < myTrips.Count() + 1; i++) {
-                foreach (Tuple<Int64, DateTime> entry in myTrips[i - 1].allTimestamps) {
+                foreach (Tuple<int, DateTime> entry in myTrips[i - 1].allTimestamps) {
                     dbc.UpdateWithNewId(i, entry.Item1);
                 }
             }
@@ -33,7 +33,7 @@ namespace CarDataProject {
             dbc.Close();
         }
 
-        public void UpdateCarWithTripIdsOptimized(Int64 carid) {
+        public void UpdateCarWithTripIdsOptimized(Int16 carid) {
             DBController dbc = new DBController();
 
             List<Trip> myTrips = TripCalculator.CalculateTripsByCarId(1);
